Remove next prefixed properties from public types
