+++ sdk/tables/data-tables/src/generatedModels.ts
@@ -8,7 +8,6 @@ export {
   TableServiceProperties as ServiceProperties,
   ServiceSetPropertiesResponse as SetPropertiesResponse,
   TableDeleteResponse as DeleteTableResponse,
-  TableQueryEntitiesWithPartitionAndRowKeyResponse,
   TableDeleteEntityResponse as DeleteTableEntityResponse,
   TableUpdateEntityResponse as UpdateEntityResponse,
   TableMergeEntityResponse as UpsertEntityResponse,
@@ -20,16 +19,12 @@ export {
   TableDeleteEntityHeaders,
   TableDeleteHeaders,
   TableGetAccessPolicyHeaders,
-  TableQueryEntitiesWithPartitionAndRowKeyHeaders,
-  TableEntityQueryResponse,
   ServiceGetPropertiesHeaders,
   ServiceGetStatisticsHeaders,
   GeoReplicationStatusType,
   KnownGeoReplicationStatusType,
   GeoReplication,
   TableServiceStats,
-  TableQueryEntitiesHeaders,
-  TableQueryHeaders,
   TableMergeEntityHeaders,
   CorsRule,
   Metrics,
